# feat: optimise for pederson hash