--- a/Cargo.toml
+++ b/Cargo.toml
@@ -92,6 +92,7 @@ env_logger = { version = "0.11.5"}
 chrono = "0.4"
 jemallocator = "0.5"
 hex = "0.4.3"
+lazy_static = "1.5.0"
 [profile.release]
 overflow-checks = true