--- a/air/src/public_memory.rs
+++ b/air/src/public_memory.rs
@@ -116,10 +116,14 @@ impl<F: SimpleField + PoseidonHash> PublicInput<F> {
         debug!("self.main_page.len() = {}", self.main_page.len());
         let current = std::time::Instant::now();
         for memory in self.main_page.iter() {
-            main_page_hash = PedersenHash::hash(main_page_hash.clone(), memory.address.clone());
-            main_page_hash = PedersenHash::hash(main_page_hash.clone(), memory.value.clone());
+            let memory = memory.clone();
+            main_page_hash = PedersenHash::hash(main_page_hash, memory.address);
+            main_page_hash = PedersenHash::hash(main_page_hash, memory.value);
         }
-        debug!("Main page hash computed in {} seconds", current.elapsed().as_secs_f32());
+        debug!(
+            "Main page hash computed in {} seconds",
+            current.elapsed().as_secs_f32()
+        );
         main_page_hash = PedersenHash::hash(
             main_page_hash,
             F::two() * F::from_constant(self.main_page.len() as u128),