--- a/fri/src/config.rs
+++ b/fri/src/config.rs
@@ -2,11 +2,11 @@ use alloc::vec::Vec;
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
 
-const MAX_LAST_LAYER_LOG_DEGREE_BOUND: u64 = 15;
-const MAX_FRI_LAYERS: u64 = 15;
-const MIN_FRI_LAYERS: u64 = 2;
-const MAX_FRI_STEP: u64 = 4;
-const MIN_FRI_STEP: u64 = 1;
+// const MAX_LAST_LAYER_LOG_DEGREE_BOUND: u64 = 15;
+// const MAX_FRI_LAYERS: u64 = 15;
+// const MIN_FRI_LAYERS: u64 = 2;
+// const MAX_FRI_STEP: u64 = 4;
+// const MIN_FRI_STEP: u64 = 1;
 
 #[serde_as]
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]