--- a/proof_parser/src/lib.rs
+++ b/proof_parser/src/lib.rs
@@ -15,6 +15,7 @@ extern crate serde;
 use ark_ff::PrimeField;
 use ark_r1cs_std::{alloc::AllocVar, fields::fp::FpVar};
 use ark_relations::r1cs::ConstraintSystem;
+use log::info;
 use swiftness_field::SimpleField;
 use swiftness_hash::poseidon::PoseidonHash;
 use swiftness_stark::types::StarkProof as StarkProofVerifier;
@@ -30,7 +31,7 @@ where
     let cs = ConstraintSystem::<F>::new_ref();
     let stark_proof_verifier: StarkProofVerifier<FpVar<F>> =
         StarkProofVerifier::<FpVar<F>>::new_witness(cs.clone(), || Ok(stark_proof)).unwrap();
-    println!("num_constraints={}", cs.num_constraints());
+    info!("num_constraints={}", cs.num_constraints());
     assert!(cs.is_satisfied().unwrap());
     Ok(stark_proof_verifier)
 }