## Problem

During testing, we observed that the line `let digest = self.public_input.get_hash();` consumes approximately 60GB of memory and takes around 160 seconds to run. This performance needs optimization.

## Work Done

Analysis showed that most of the time is spent in the function `gen_element_steps_var`, which takes about 150 seconds. We optimized it by:

	1.	Using global variables to replace constants, avoiding redundant calculations.
	2.	Improving bitwise operation efficiency.

## Result

The optimization was successful, reducing memory usage from 60GB to 3GB and execution time from 160 seconds to 27 seconds.